AsyncQueue/requirements.txt
websockets
colorama
PyYAML
fastapi
spacy
parsedatetime
requests
pandas
tweepy
textblob